+++ src/backend.rs
@@ -1,21 +1,21 @@
-use crate::scripts::MintingPolicy;
-use crate::values::Values;
 use crate::{
     address::{Address, PolicyId},
-    error::Error,
+    backend::nested_value_map::{add_amount_to_nested_map, nested_map_to_vecs},
+    backend::tx_checks::{can_mint_tokens, can_spend_inputs},
     error::Result,
     ledger_client::LedgerClient,
     output::Output,
-    scripts::{TxContext, ValidatorCode},
+    scripts::MintingPolicy,
     transaction::Action,
+    values::Values,
     Transaction, UnBuiltTransaction,
 };
-use std::borrow::Borrow;
-use std::cmp::min;
 use std::{cell::RefCell, collections::HashMap, fmt::Debug, hash::Hash, marker::PhantomData};
 use uuid::Uuid;
 
+mod nested_value_map;
 pub mod selection;
+pub mod tx_checks;
 
 #[cfg(test)]
 mod tests;
@@ -63,23 +63,16 @@ where
     fn handle_actions(
         &self,
         actions: Vec<Action<Datum, Redeemer>>,
-    ) -> Result<(
-        Vec<Output<Datum>>,
-        Vec<Output<Datum>>,
-        Vec<(Output<Datum>, Redeemer)>,
-        HashMap<Address, Box<dyn ValidatorCode<Datum, Redeemer>>>,
-        HashMap<PolicyId, u64>,
-        HashMap<Address, Box<dyn MintingPolicy>>,
-    )> {
+    ) -> Result<Transaction<Datum, Redeemer>> {
         let mut min_input_values = Values::default();
         let mut min_output_values: HashMap<Address, RefCell<Values>> = HashMap::new();
-        let mut minting: HashMap<PolicyId, u64> = HashMap::new();
+        let mut minting = Values::default();
         let mut script_inputs: Vec<Output<Datum>> = Vec::new();
         let mut specific_outputs: Vec<Output<Datum>> = Vec::new();
 
         let mut redeemers = Vec::new();
-        let mut validator_scripts = HashMap::new();
-        let mut policy_scripts: HashMap<Address, Box<dyn MintingPolicy>> = HashMap::new();
+        let mut validators = HashMap::new();
+        let mut policies: HashMap<Address, Box<dyn MintingPolicy>> = HashMap::new();
         for action in actions {
             match action {
                 Action::Transfer {
@@ -88,7 +81,6 @@ where
                     policy_id: policy,
                 } => {
                     // Input
-                    // add_to_map(&mut min_input_values, policy.clone(), amount);
                     min_input_values.add_one_value(&policy, amount);
 
                     // Output
@@ -106,8 +98,8 @@ where
                         &recipient,
                         &policy_id,
                     );
-                    add_to_map(&mut minting, policy_id.clone(), amount);
-                    policy_scripts.insert(policy.address(), policy);
+                    minting.add_one_value(&policy_id, amount);
+                    policies.insert(policy.address(), policy);
                 }
                 Action::InitScript {
                     datum,
@@ -115,8 +107,7 @@ where
                     address,
                 } => {
                     for (policy, amount) in values.iter() {
-                        // add_to_map(&mut min_input_values, policy.clone(), *amount);
-                        min_input_values.add_one_value(&policy, *amount);
+                        min_input_values.add_one_value(policy, *amount);
                     }
                     let id = Uuid::new_v4().to_string(); // TODO: This should be done by the TxORecord impl or something
                     let owner = address;
@@ -136,7 +127,7 @@ where
                     script_inputs.push(output.clone());
                     let script_address = script.address();
                     redeemers.push((output, redeemer));
-                    validator_scripts.insert(script_address, script);
+                    validators.insert(script_address, script);
                 }
             }
         }
@@ -144,19 +135,9 @@ where
         let (inputs, remainders) =
             self.select_inputs_for_one(self.txo_record.signer(), &min_input_values, script_inputs)?;
 
-        // outputs
-        // remainders.iter().for_each(|(amt, policy)| {
-        //     add_amount_to_nested_map(
-        //         &mut min_output_values,
-        //         *amt,
-        //         &self.txo_record.signer(),
-        //         policy,
-        //     )
-        // });
-
         // TODO: Dedupe
         let mut new_values = remainders;
-        if let Some(values) = min_output_values.get(&self.txo_record.signer()) {
+        if let Some(values) = min_output_values.get(self.txo_record.signer()) {
             new_values.add_values(&values.borrow());
         }
         min_output_values.insert(self.txo_record.signer().clone(), RefCell::new(new_values));
@@ -165,14 +146,15 @@ where
         let mut outputs = self.create_outputs_for(out_vecs)?;
         outputs.extend(specific_outputs);
 
-        Ok((
+        let tx = Transaction {
             inputs,
             outputs,
             redeemers,
-            validator_scripts,
+            validators,
             minting,
-            policy_scripts,
-        ))
+            policies,
+        };
+        Ok(tx)
     }
 
     // LOL Super Naive Solution, just select ALL inputs!
@@ -183,7 +165,6 @@ where
     fn select_inputs_for_one(
         &self,
         address: &Address,
-        // values: &HashMap<PolicyId, u64>,
         spending_values: &Values,
         script_inputs: Vec<Output<Datum>>,
     ) -> Result<(Vec<Output<Datum>>, Values)> {
@@ -216,104 +197,7 @@ where
         unbuilt_tx: UnBuiltTransaction<Datum, Redeemer>,
     ) -> Result<Transaction<Datum, Redeemer>> {
         let UnBuiltTransaction { actions } = unbuilt_tx;
-        let (inputs, outputs, redeemers, scripts, minting, policies) =
-            self.handle_actions(actions)?;
+        self.handle_actions(actions)
         // TODO: Calculate fees and then rebuild tx
-        Ok(Transaction {
-            inputs,
-            outputs,
-            redeemers,
-            scripts,
-            minting,
-            policies,
-        })
     }
-}
-
-pub fn add_to_map(h_map: &mut HashMap<PolicyId, u64>, policy: PolicyId, amount: u64) {
-    let mut new_total = amount;
-    if let Some(total) = h_map.get(&policy) {
-        new_total += total;
-    }
-    h_map.insert(policy.clone(), new_total);
-}
-
-fn nested_map_to_vecs(
-    nested_map: HashMap<Address, RefCell<Values>>,
-) -> Vec<(Address, Vec<(PolicyId, u64)>)> {
-    nested_map
-        .into_iter()
-        .map(|(addr, values)| (addr, values.borrow().vec()))
-        .collect()
-}
-
-fn add_amount_to_nested_map(
-    output_map: &mut HashMap<Address, RefCell<Values>>,
-    amount: u64,
-    owner: &Address,
-    policy_id: &PolicyId,
-) {
-    if let Some(values) = output_map.get(owner) {
-        let mut inner = values.borrow_mut();
-        // let mut new_total = amount;
-        // if let Some(total) = inner.get(policy_id) {
-        //     new_total += total;
-        // }
-        // inner.insert(policy_id.clone(), new_total);
-        inner.add_one_value(policy_id, amount);
-    } else {
-        let mut new_values = Values::default();
-        new_values.add_one_value(&policy_id, amount);
-        output_map.insert(owner.clone(), RefCell::new(new_values));
-    }
-}
-
-pub fn can_spend_inputs<
-    Datum: Clone + PartialEq + Debug,
-    Redeemer: Clone + PartialEq + Eq + Hash,
->(
-    tx: &Transaction<Datum, Redeemer>,
-    signer: Address,
-) -> Result<()> {
-    let ctx = TxContext { signer };
-    for input in &tx.inputs {
-        match input {
-            Output::Wallet { .. } => {} // TODO: Make sure not spending other's outputs
-            Output::Validator { owner, datum, .. } => {
-                let script = tx
-                    .scripts
-                    .get(owner)
-                    .ok_or_else(|| Error::FailedToRetrieveScriptFor(owner.to_owned()))?;
-                let (_, redeemer) = tx
-                    .redeemers
-                    .iter()
-                    .find(|(utxo, _)| utxo == input)
-                    .ok_or_else(|| Error::FailedToRetrieveRedeemerFor(owner.to_owned()))?;
-
-                script.execute(datum.clone(), redeemer.clone(), ctx.clone())?;
-            }
-        }
-    }
-    Ok(())
-}
-
-pub fn can_mint_tokens<Datum, Redeemer>(
-    tx: &Transaction<Datum, Redeemer>,
-    signer: &Address,
-) -> Result<()> {
-    let ctx = TxContext {
-        signer: signer.clone(),
-    };
-    for (id, _) in tx.minting.iter() {
-        if let Some(address) = id {
-            if let Some(policy) = tx.policies.get(address) {
-                policy.execute(ctx.clone())?;
-            } else {
-                return Err(Error::FailedToRetrieveScriptFor(address.clone()));
-            }
-        } else {
-            return Err(Error::ImpossibleToMintADA);
-        }
-    }
-    Ok(())
 }
