+++ src/values.rs
@@ -1,9 +1,7 @@
-use crate::backend::add_to_map;
 use crate::error::{Error, Result};
 use crate::output::Output;
 use crate::PolicyId;
 use std::collections::HashMap;
-use std::ops::{Add, Sub};
 
 #[derive(Default, Clone)]
 pub struct Values {
@@ -14,8 +12,7 @@ impl Values {
     pub fn from_outputs<D>(outputs: &[Output<D>]) -> Self {
         let mut values = HashMap::new();
         outputs
-            .clone()
-            .into_iter()
+            .iter()
             .flat_map(|o| o.values().clone().into_iter().collect::<Vec<_>>())
             .for_each(|(policy, amount)| {
                 add_to_map(&mut values, policy, amount);
@@ -23,15 +20,10 @@ impl Values {
         Values { values }
     }
 
-    pub fn add_output_value<D>(&mut self, output: &Output<D>) {
-        todo!()
-    }
-
     pub fn try_subtract(&self, other: &Values) -> Result<Values> {
-        // TODO: make more efficient
         let mut remainders = Vec::new();
         let mut mine_cloned = self.values.clone();
-        for (policy, amt) in other.vec().iter() {
+        for (policy, amt) in other.as_iter() {
             if let Some(available) = mine_cloned.remove(policy) {
                 if amt <= &available {
                     let remaining = available - amt;
@@ -54,16 +46,16 @@ impl Values {
     pub fn add_one_value(&mut self, policy: &PolicyId, amount: u64) {
         add_to_map(&mut self.values, policy.clone(), amount)
     }
-    //
-    // pub fn remove_one_policy(&mut self, policy: &PolicyId) -> Option<u64> {
-    //     self.values.remove(policy)
-    // }
 
     pub fn add_values(&mut self, values: &Values) {
         // TODO: make more efficient
-        for (policy, amt) in values.vec().iter() {
+        for (policy, amt) in values.as_iter() {
             self.add_one_value(policy, *amt)
         }
+    }
+
+    pub fn as_iter(&self) -> std::collections::hash_map::Iter<'_, PolicyId, u64> {
+        self.values.iter()
     }
 
     pub fn vec(&self) -> Vec<(PolicyId, u64)> {
@@ -71,10 +63,10 @@ impl Values {
     }
 }
 
-impl Add for Values {
-    type Output = Self;
-
-    fn add(self, rhs: Self) -> Self::Output {
-        todo!()
+pub fn add_to_map(h_map: &mut HashMap<PolicyId, u64>, policy: PolicyId, amount: u64) {
+    let mut new_total = amount;
+    if let Some(total) = h_map.get(&policy) {
+        new_total += total;
     }
+    h_map.insert(policy.clone(), new_total);
 }
